+++ NEWS.md
@@ -1,6 +1,7 @@
 # tmap 4.4 (in development)
 
 - [!] added blend argument to all layers
+- [!] added tm_circles (circles with fixed unit-based radii)
 - [!] added hitbox to tm_lines and tm_symbols for view mode (so hovering and clicking also works well for small objects)
 - added `tmap_mode_pool()` to restrict which modes are cycled by `ttm()` and `rtm()`, e.g. `library(tmap.mapgl); tmap_mode_pool(c("plot", "maplibre"))`
 - shiny dispatch methods added (to allow integration for other modes)